Feat: Post 화면에서 보이는 헤더 코드 스플릿 진행
