+++ client/src/components/UI/organisms/PostHeader/postheader.module.scss
@@ -1,0 +1,30 @@
+@import '/src/utils/utils';
+
+header {
+  width: 100%;
+  height: #{70 / $font-size}rem;
+  position: fixed;
+  background-color: $white;
+  border: 1px solid $gray-white;
+  z-index: 2;
+
+  nav {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    padding: 0 2rem;
+    justify-content: space-between;
+    align-items: center;
+
+    .upload {
+      display: flex;
+      align-items: center;
+      width: 100%;
+      justify-content: space-evenly;
+    }
+    .save {
+      display: flex;
+      align-items: center;
+    }
+  }
+}
