+++ client/src/components/UI/organisms/UserAuth/UserAuth.tsx
@@ -12,7 +12,7 @@ import {
 } from '../../../../utils/utils';
 import { useNavigate } from 'react-router-dom';
 import { userInfo } from '../../../../redux/UserSlice';
-import { toast } from 'react-toastify';
+
 interface UserAuthProps {
   type: 'Login' | 'SignUp';
 }
@@ -30,6 +30,7 @@ const UserAuth = ({ type }: UserAuthProps) => {
       await handleUserLogin(userLoginData.email, userLoginData.password).then(
         (response) => {
           if (typeof response !== 'undefined') {
+            console.log(response);
             const { displayName, email, phoneNumber, photoURL } = response;
             dispatch(userInfo({ displayName, email, phoneNumber, photoURL }));
           }
@@ -41,23 +42,13 @@ const UserAuth = ({ type }: UserAuthProps) => {
         userRegData.email,
         userRegData.password,
         userRegData.nickname
-      ).then((response) => {
-        if (typeof response !== 'undefined') {
-          const { displayName, email, phoneNumber, photoURL } = response;
-          dispatch(userInfo({ displayName, email, phoneNumber, photoURL }));
-        }
-      });
+      );
       navigate('/');
     } else null;
   };
 
   const handleOauth = async () => {
-    await handleGoogleLogin().then((response) => {
-      if (typeof response !== 'undefined') {
-        const { displayName, email, phoneNumber, photoURL } = response;
-        dispatch(userInfo({ displayName, email, phoneNumber, photoURL }));
-      }
-    });
+    await handleGoogleLogin();
     navigate('/');
   };
