+++ client/src/components/pages/Main/Main.tsx
@@ -34,7 +34,7 @@ function Main() {
   const [musicArr, setMusicArr] = useState<MusicData>([
     { artist: '', scores: [], albumImg: '', songName: '', songId: null },
   ]);
-  console.log(musicArr);
+
   useEffect(() => {
     getMusics().then((data) => {
       setMusicArr(data);
