+++ client/src/components/pages/PostLayout/postlayout.module.scss
@@ -1,0 +1,5 @@
+@import '/src/utils/utils';
+
+main {
+  padding-top: #{70 / $font-size}rem;
+}
